move constants.ts to src
