+++ src/utils/gql.ts
@@ -1,4 +1,4 @@
-import {GRAPH_URL} from './constants';
+import {GRAPH_URL} from '../constants';
 
 export async function request(query: string, variables = {}, url = GRAPH_URL) {
   const _ = await fetch(url, {
